chore: activate stricNullCheck

+++ src/core/cast-member/infra/db/sequelize/migrations/2024.09.27T02.21.12.create-cast-members-table.ts
@@ -13,7 +13,7 @@ export const up: MigrationFn<Sequelize> = async ({ context: sequelize }) => {
       allowNull: false,
     },
     type: {
-      type: DataType.INTEGER,
+      type: DataType.SMALLINT,
       allowNull: false,
     },
     created_at: {
